FM-39: Move arbitraries to the testing utils.
