+++ fendermint/testing/src/lib.rs
@@ -1,4 +1,8 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
+#[cfg(feature = "arb")]
+pub mod arb;
+#[cfg(feature = "golden")]
 pub mod golden;
+#[cfg(feature = "smt")]
 pub mod smt;
