+++ fendermint/vm/message/src/chain.rs
@@ -35,15 +35,17 @@ pub enum ChainMessage {
 
 #[cfg(feature = "arb")]
 mod arb {
+    use fendermint_testing::arb::ArbCid;
+
     use super::ChainMessage;
-    use crate::{arb::cid::arbitrary_cid, signed::SignedMessage};
+    use crate::signed::SignedMessage;
 
     impl quickcheck::Arbitrary for ChainMessage {
         fn arbitrary(g: &mut quickcheck::Gen) -> Self {
             match u8::arbitrary(g) % 3 {
                 0 => ChainMessage::Signed(Box::new(SignedMessage::arbitrary(g))),
-                1 => ChainMessage::ForResolution(arbitrary_cid(g)),
-                _ => ChainMessage::ForExecution(arbitrary_cid(g)),
+                1 => ChainMessage::ForResolution(ArbCid::arbitrary(g).0),
+                _ => ChainMessage::ForExecution(ArbCid::arbitrary(g).0),
             }
         }
     }
