+++ fendermint/vm/message/src/lib.rs
@@ -4,9 +4,6 @@ use cid::{multihash, multihash::MultihashDigest, Cid};
 use fvm_ipld_encoding::{to_vec, Error as IpldError, DAG_CBOR};
 use serde::Serialize;
 
-#[cfg(feature = "arb")]
-mod arb;
-
 pub mod chain;
 pub mod query;
 pub mod signed;
